Updated Bazel
