+++ Client/Client.cpp
@@ -4,7 +4,6 @@
 
 #include <iostream>
 #include <string>
-#include "..\Communication\Create.h"
 #include "Client.h"
