+++ ClientCLI/ClientCLI.h
@@ -1,6 +1,4 @@
 #pragma once
-#include "IContainer.h"
-
 #include "..\Communication\Create.h"
 #include "..\Logger\Create.h"
 #include "..\Server\Create.h"
