+++ Logger/Logger.h
@@ -1,7 +1,7 @@
 #pragma once
-#include "IContainer.h"
 #include <ctime>
 #include <vector>
+#include "../Communication/Create.h"
 
 namespace Logger {
 	enum action {
