+++ Tamagotchi/Create.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "IContainer.h"
 #include "Tamagotchi.h"
 #include <memory>
