+++ Tamagotchi/Tamagotchi.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "IContainer.h"
 #include <cstdint>
 #include <deque>
 #include "../Communication/Create.h"
