+++ Testing/Mocks/Communication.cpp
@@ -1,0 +1,1 @@
+#include "Communication.h"
